Conflicts:
	avengers/src/main/java/com/flyhz/avengers/common/event/URLCrawlEvent.java
	avengers/src/main/java/com/flyhz/avengers/framework/Crawl.java
	avengers/src/main/java/com/flyhz/avengers/framework/config/XConfiguration.java
	avengers/src/main/resources/avengers.xml
 Changes to be committed:
 	modified:   src/main/java/com/flyhz/avengers/common/event/TemplateApplyEvent.java
 	modified:   src/main/java/com/flyhz/avengers/common/event/URLCrawlEvent.java
 	modified:   src/main/java/com/flyhz/avengers/common/event/URLFetchEvent.java
 	new file:   src/main/java/com/flyhz/avengers/domains/abercrombie/template/AbercrombieProductTemplateImpl.java
 	modified:   src/main/java/com/flyhz/avengers/framework/Analyze.java
 	modified:   src/main/java/com/flyhz/avengers/framework/AvengersAppMaster.java
 	modified:   src/main/java/com/flyhz/avengers/framework/Crawl.java
 	modified:   src/main/java/com/flyhz/avengers/framework/Fetch.java
 	modified:   src/main/java/com/flyhz/avengers/framework/Template.java
 	modified:   src/main/java/com/flyhz/avengers/framework/config/XConfiguration.java
 	modified:   src/main/java/com/flyhz/avengers/framework/config/xml/ObjectFactory.java
 	modified:   src/main/java/com/flyhz/avengers/framework/config/xml/XDomain.java
 	new file:   src/main/java/com/flyhz/avengers/framework/util/SerializableUtil.java
 	modified:   src/main/java/com/flyhz/avengers/framework/util/StringUtil.java
 	deleted:    src/main/java/com/flyhz/avengers/framework/util/filter/URLFilterUtil.java
 	modified:   src/main/resources/avengers.xml
 	modified:   src/main/resources/avengers.xsd
 	modified:   ../smile_android/gen/com/holding/smile/R.java
 	modified:   ../smile_android/jni/m.cpp
 	modified:   ../smile_android/res/layout/good_pic_item.xml
 	modified:   ../smile_android/res/layout/index_item.xml
 	modified:   ../smile_android/res/layout/smile_main.xml
 	modified:   ../smile_android/res/menu/main.xml
 	modified:   ../smile_android/src/com/holding/smile/activity/BaseActivity.java
 	modified:   ../smile_android/src/com/holding/smile/activity/CategoryActivity.java
 	modified:   ../smile_android/src/com/holding/smile/activity/GoodsDetailActivity.java
 	modified:   ../smile_android/src/com/holding/smile/activity/IdcardEditActivity.java
 	modified:   ../smile_android/src/com/holding/smile/activity/IdcardManagerActivity.java
 	modified:   ../smile_android/src/com/holding/smile/activity/LoginActivity.java
 	modified:   ../smile_android/src/com/holding/smile/activity/MainSmileActivity.java
 	modified:   ../smile_android/src/com/holding/smile/activity/MainTwoActivity.java
 	modified:   ../smile_android/src/com/holding/smile/activity/MyApplication.java
 	modified:   ../smile_android/src/com/holding/smile/activity/MySmileActivity.java
 	modified:   ../smile_android/src/com/holding/smile/activity/OrderInformActivity.java
 	modified:   ../smile_android/src/com/holding/smile/activity/SearchGoodsActivity.java
 	modified:   ../smile_android/src/com/holding/smile/activity/ShoppingCartActivity.java
 	modified:   ../smile_android/src/com/holding/smile/activity/SortActivity.java
 	modified:   ../smile_android/src/com/holding/smile/activity/WebViewActivity.java
 	modified:   ../smile_android/src/com/holding/smile/adapter/BrandAdapter.java

+++ avengers/src/main/java/com/flyhz/avengers/framework/config/xml/XDomain.java
@@ -30,13 +30,15 @@ import javax.xml.bind.annotation.XmlType;
  *       &lt;attribute name="encoding" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" default="UTF-8" />
  *       &lt;attribute name="depth" default="3">
  *         &lt;simpleType>
- *           &lt;restriction base="{http://www.w3.org/2001/XMLSchema}unsignedInt">
+ *           &lt;restriction base="{http://www.w3.org/2001/XMLSchema}unsignedByte">
+ *             &lt;maxExclusive value="15"/>
  *           &lt;/restriction>
  *         &lt;/simpleType>
  *       &lt;/attribute>
  *       &lt;attribute name="period" default="720">
  *         &lt;simpleType>
- *           &lt;restriction base="{http://www.w3.org/2001/XMLSchema}unsignedInt">
+ *           &lt;restriction base="{http://www.w3.org/2001/XMLSchema}unsignedShort">
+ *             &lt;maxExclusive value="744"/>
  *           &lt;/restriction>
  *         &lt;/simpleType>
  *       &lt;/attribute>
@@ -65,9 +67,9 @@ public class XDomain {
 	@XmlSchemaType(name = "anySimpleType")
 	protected String		encoding;
 	@XmlAttribute
-	protected Long			depth;
+	protected Short			depth;
 	@XmlAttribute
-	protected Long			period;
+	protected Integer		period;
 
 	/**
 	 * Gets the value of the urlFilterBeforeCrawl property.
@@ -223,12 +225,12 @@ public class XDomain {
 	/**
 	 * Gets the value of the depth property.
 	 * 
-	 * @return possible object is {@link Long }
+	 * @return possible object is {@link Short }
 	 * 
 	 */
-	public long getDepth() {
+	public short getDepth() {
 		if (depth == null) {
-			return 3L;
+			return ((short) 3);
 		} else {
 			return depth;
 		}
@@ -238,22 +240,22 @@ public class XDomain {
 	 * Sets the value of the depth property.
 	 * 
 	 * @param value
-	 *            allowed object is {@link Long }
+	 *            allowed object is {@link Short }
 	 * 
 	 */
-	public void setDepth(Long value) {
+	public void setDepth(Short value) {
 		this.depth = value;
 	}
 
 	/**
 	 * Gets the value of the period property.
 	 * 
-	 * @return possible object is {@link Long }
+	 * @return possible object is {@link Integer }
 	 * 
 	 */
-	public long getPeriod() {
+	public int getPeriod() {
 		if (period == null) {
-			return 720L;
+			return 720;
 		} else {
 			return period;
 		}
@@ -263,10 +265,10 @@ public class XDomain {
 	 * Sets the value of the period property.
 	 * 
 	 * @param value
-	 *            allowed object is {@link Long }
+	 *            allowed object is {@link Integer }
 	 * 
 	 */
-	public void setPeriod(Long value) {
+	public void setPeriod(Integer value) {
 		this.period = value;
 	}
